Move UriSigner from HttpKernel to HttpFoundation package
