+++ src/Symfony/Bridge/Twig/composer.json
@@ -31,7 +31,7 @@
         "symfony/form": "^6.3|^7.0",
         "symfony/html-sanitizer": "^6.1|^7.0",
         "symfony/http-foundation": "^5.4|^6.0|^7.0",
-        "symfony/http-kernel": "^6.2|^7.0",
+        "symfony/http-kernel": "^6.4|^7.0",
         "symfony/intl": "^5.4|^6.0|^7.0",
         "symfony/mime": "^6.2|^7.0",
         "symfony/polyfill-intl-icu": "~1.0",
@@ -59,7 +59,7 @@
         "symfony/console": "<5.4",
         "symfony/form": "<6.3",
         "symfony/http-foundation": "<5.4",
-        "symfony/http-kernel": "<6.2",
+        "symfony/http-kernel": "<6.4",
         "symfony/mime": "<6.2",
         "symfony/serializer": "<6.2",
         "symfony/translation": "<5.4",
