+++ src/Symfony/Component/HttpFoundation/Tests/UriSignerTest.php
@@ -9,15 +9,12 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\HttpKernel\Tests;
+namespace Symfony\Component\HttpFoundation\Tests;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpKernel\UriSigner;
+use Symfony\Component\HttpFoundation\UriSigner;
 
-/**
- * @group legacy
- */
 class UriSignerTest extends TestCase
 {
     public function testSign()
