+++ src/Symfony/Component/HttpKernel/composer.json
@@ -20,7 +20,7 @@
         "symfony/deprecation-contracts": "^2.5|^3",
         "symfony/error-handler": "^6.4|^7.0",
         "symfony/event-dispatcher": "^5.4|^6.0|^7.0",
-        "symfony/http-foundation": "^6.3.4|^7.0",
+        "symfony/http-foundation": "^6.4|^7.0",
         "symfony/polyfill-ctype": "^1.8",
         "psr/log": "^1|^2|^3"
     },
